move zip handlers to reader directory in unicorn
