+++ src/allotropy/parsers/cytiva_unicorn/cytiva_unicorn_parser.py
@@ -7,7 +7,9 @@ from allotropy.allotrope.schema_mappers.adm.liquid_chromatography.benchling._202
 )
 from allotropy.named_file_contents import NamedFileContents
 from allotropy.parsers.cytiva_unicorn.constants import DISPLAY_NAME
-from allotropy.parsers.cytiva_unicorn.cytiva_unicorn_reader import UnicornFileHandler
+from allotropy.parsers.cytiva_unicorn.reader.unicorn_zip_handler import (
+    UnicornZipHandler,
+)
 from allotropy.parsers.cytiva_unicorn.structure.measurement_group import (
     create_measurement_groups,
 )
@@ -25,7 +27,7 @@ class CytivaUnicornParser(VendorParser[Data, Model]):
     SCHEMA_MAPPER = Mapper
 
     def create_data(self, named_file_contents: NamedFileContents) -> Data:
-        handler = UnicornFileHandler(named_file_contents.original_file_path)
+        handler = UnicornZipHandler(named_file_contents.original_file_path)
         results = handler.get_results()
         return Data(
             create_metadata(handler, results),
