+++ src/allotropy/parsers/cytiva_unicorn/reader/unicorn_zip_handler.py
@@ -1,50 +1,15 @@
-from __future__ import annotations
-
-from io import BytesIO
 from re import search
-from zipfile import Path, ZipFile
 
 from allotropy.exceptions import AllotropeConversionError
 from allotropy.parsers.cytiva_unicorn.reader.strict_element import (
     StrictElement,
 )
-from allotropy.parsers.utils.values import assert_not_none
+from allotropy.parsers.cytiva_unicorn.reader.zip_handler import (
+    ZipHandler,
+)
 
 
-class ZipHandler:
-    def __init__(self, path: str, inner_path: str = "", data: BytesIO | None = None):
-        self.path = path
-        self.inner_path = inner_path
-        self.zip_path = Path(path, inner_path)
-        self.zip_file = ZipFile(path if data is None else data)
-        self.name_list = self.zip_file.namelist()
-
-    def get_file_path_or_none(self, pattern: str) -> str | None:
-        for file_name in self.name_list:
-            if search(pattern, file_name):
-                return file_name
-        return None
-
-    def get_file_path(self, pattern: str) -> str:
-        return assert_not_none(
-            self.get_file_path_or_none(pattern),
-            msg=f"Unable to find file inside {self.path} that match pattern {pattern}",
-        )
-
-    def __get_raw_content(self, inner_path: str) -> BytesIO:
-        return BytesIO(self.zip_file.read(inner_path))
-
-    def __get_content(self, inner_path: str) -> ZipHandler:
-        return ZipHandler(self.path, inner_path, self.__get_raw_content(inner_path))
-
-    def get_file_from_pattern(self, pattern: str) -> BytesIO:
-        return self.__get_raw_content(self.get_file_path(pattern))
-
-    def get_content_from_pattern(self, pattern: str) -> ZipHandler:
-        return self.__get_content(self.get_file_path(pattern))
-
-
-class UnicornFileHandler(ZipHandler):
+class UnicornZipHandler(ZipHandler):
     def get_system_data(self) -> StrictElement:
         system_data = self.get_content_from_pattern("SystemData.zip$")
         b_stream = system_data.get_file_from_pattern("^Xml$")
