+++ src/allotropy/parsers/cytiva_unicorn/structure/measurements/ph.py
@@ -7,11 +7,11 @@ from allotropy.allotrope.schema_mappers.adm.liquid_chromatography.benchling._202
     Measurement,
 )
 from allotropy.parsers.cytiva_unicorn.constants import DEVICE_TYPE
-from allotropy.parsers.cytiva_unicorn.cytiva_unicorn_reader import (
-    UnicornFileHandler,
-)
 from allotropy.parsers.cytiva_unicorn.reader.strict_element import (
     StrictElement,
+)
+from allotropy.parsers.cytiva_unicorn.reader.unicorn_zip_handler import (
+    UnicornZipHandler,
 )
 from allotropy.parsers.cytiva_unicorn.structure.measurements.generic import (
     UnicornMeasurement,
@@ -26,7 +26,7 @@ class PhMeasurement(UnicornMeasurement):
     @classmethod
     def create(
         cls,
-        handler: UnicornFileHandler,
+        handler: UnicornZipHandler,
         elements: list[StrictElement],
         static_docs: StaticDocs,
     ) -> Measurement:
