+++ src/allotropy/parsers/cytiva_unicorn/structure/measurements/pressure.py
@@ -8,11 +8,11 @@ from allotropy.allotrope.schema_mappers.adm.liquid_chromatography.benchling._202
     ProcessedDataDoc,
 )
 from allotropy.parsers.cytiva_unicorn.constants import DEVICE_TYPE
-from allotropy.parsers.cytiva_unicorn.cytiva_unicorn_reader import (
-    UnicornFileHandler,
-)
 from allotropy.parsers.cytiva_unicorn.reader.strict_element import (
     StrictElement,
+)
+from allotropy.parsers.cytiva_unicorn.reader.unicorn_zip_handler import (
+    UnicornZipHandler,
 )
 from allotropy.parsers.cytiva_unicorn.structure.measurements.generic import (
     UnicornMeasurement,
@@ -27,7 +27,7 @@ class PressureMeasurement(UnicornMeasurement):
     @classmethod
     def create(
         cls,
-        handler: UnicornFileHandler,
+        handler: UnicornZipHandler,
         elements: list[StrictElement],
         static_docs: StaticDocs,
     ) -> Measurement:
